requirements.txt
tensorflow
numpy
matplotlib
opencv-python
streamlit
gradio
Pillow